Refactored separate props to Friend interface
