+++ src/types/Friend.d.ts
@@ -1,0 +1,6 @@
+export default interface Friend {
+  id: number;
+  name: string;
+  image: string;
+  balance: number;
+}
